requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
elastic-transport==8.13.1
elasticsearch==8.14.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.24
langchain-openai==0.1.19
langsmith==0.1.93
numpy==2.0.1
openai==1.37.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
